chore: dependency cleanup
