+++ src/notebook_provider/constants.ts
@@ -1,0 +1,2 @@
+export const SupportedLanguages = ["python", "sql", "jinja-sql"];
+export const DefaultNotebookCellLanguage = "jinja-sql";
